client/requirements.txt
grpcio==1.12.1
grpcio-tools==1.12.1